Remove Angelscript from Utility
