+++ src/Utility/CMakeLists.txt
@@ -9,6 +9,6 @@ set(HEADERS
 create_directory_groups(${SRCS} ${HEADERS})
 
 add_library(Utility STATIC ${SRCS} ${HEADERS})
-target_link_libraries(Utility angelscript glew glfw ${GLFW_LIBRARIES} glm)
+target_link_libraries(Utility glew glfw ${GLFW_LIBRARIES} glm)
 set_property(TARGET Utility PROPERTY CXX_STANDARD 11)
 set_property(TARGET Utility PROPERTY CXX_STANDARD_REQUIRED ON)
